# web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Walrus.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/stats": {
      "get": {
        "tags": [
          "Endpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sinks": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SinkResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/sinks/{id}": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SinkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/sinks/{id}/dead-letters": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeadLetterResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/sinks/{id}/search": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/conflicts": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/events": {
      "get": {
        "tags": [
          "Walrus.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseItemOfFeedEntry"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sinks/{id}/replay": {
      "post": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromLsn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reset",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/sinks/{id}/snapshot": {
      "post": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/sinks/{id}/dead-letters/retry": {
      "post": {
        "tags": [
          "Walrus.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangeOperation": {
        "enum": [
          "Insert",
          "Update",
          "Delete"
        ]
      },
      "ColumnValue": {
        "required": [
          "name",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DeadLetterResponse": {
        "required": [
          "id",
          "source",
          "seq",
          "table",
          "operation",
          "key",
          "error",
          "attempts",
          "deadAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "source": {
            "type": "string"
          },
          "seq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "table": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/ChangeOperation"
          },
          "key": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnValue"
            }
          },
          "error": {
            "type": "string"
          },
          "attempts": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "deadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FeedEntry": {
        "required": [
          "kind",
          "at",
          "source",
          "table",
          "operation",
          "key",
          "commitLsn",
          "sink",
          "detail"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/FeedKind"
          },
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "table": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "commitLsn": {
            "type": "string"
          },
          "sink": {
            "type": [
              "null",
              "string"
            ]
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "FeedKind": {
        "enum": [
          "Change",
          "Conflict",
          "DeadLetter"
        ]
      },
      "RegisterSinkRequest": {
        "required": [
          "id",
          "kind",
          "tables",
          "target",
          "start",
          "startLsn"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SinkKind"
          },
          "tables": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "target": {
            "type": [
              "null",
              "string"
            ]
          },
          "start": {
            "$ref": "#/components/schemas/SinkStartMode"
          },
          "startLsn": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RetryResponse": {
        "required": [
          "resolved",
          "stillBlocked"
        ],
        "type": "object",
        "properties": {
          "resolved": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "stillBlocked": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SearchHit": {
        "required": [
          "table",
          "key",
          "row"
        ],
        "type": "object",
        "properties": {
          "table": {
            "type": "string"
          },
          "key": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnValue"
            }
          },
          "row": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnValue"
            }
          }
        }
      },
      "SearchResponse": {
        "required": [
          "rows"
        ],
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            }
          }
        }
      },
      "SinkCountersResponse": {
        "required": [
          "changed",
          "unchanged",
          "conflicts",
          "deadLetters",
          "retries",
          "lastError"
        ],
        "type": "object",
        "properties": {
          "changed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "unchanged": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "conflicts": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "deadLetters": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "retries": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "lastError": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SinkKind": {
        "enum": [
          "Postgres",
          "Index",
          "Webhook"
        ]
      },
      "SinkResponse": {
        "required": [
          "id",
          "kind",
          "tables",
          "target",
          "state",
          "blockedRows",
          "sources",
          "counters"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SinkKind"
          },
          "tables": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "target": {
            "type": [
              "null",
              "string"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/SinkState"
          },
          "blockedRows": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SinkSourceResponse"
            }
          },
          "counters": {
            "$ref": "#/components/schemas/SinkCountersResponse"
          }
        }
      },
      "SinkSourceResponse": {
        "required": [
          "source",
          "appliedSeq",
          "headSeq",
          "startLsn",
          "pendingLsn",
          "lagMilliseconds"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "appliedSeq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "headSeq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "startLsn": {
            "type": "string"
          },
          "pendingLsn": {
            "type": [
              "null",
              "string"
            ]
          },
          "lagMilliseconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "SinkStartMode": {
        "enum": [
          "Beginning",
          "Now",
          "Lsn",
          "Snapshot"
        ]
      },
      "SinkState": {
        "enum": [
          "Running",
          "Retrying",
          "Rebuilding",
          "Stopped"
        ]
      },
      "SourceResponse": {
        "required": [
          "name",
          "attachedTo",
          "primary",
          "walRetainedBytes",
          "confirmedFlushLsn",
          "slotActive",
          "standbys",
          "captureIsSynchronous",
          "changes",
          "transactions",
          "resent",
          "lastCommit",
          "lastError",
          "checkedAt"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "attachedTo": {
            "type": [
              "null",
              "string"
            ]
          },
          "primary": {
            "type": [
              "null",
              "string"
            ]
          },
          "walRetainedBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "confirmedFlushLsn": {
            "type": [
              "null",
              "string"
            ]
          },
          "slotActive": {
            "type": "boolean"
          },
          "standbys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StandbyResponse"
            }
          },
          "captureIsSynchronous": {
            "type": "boolean"
          },
          "changes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "transactions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "resent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "lastCommit": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastError": {
            "type": [
              "null",
              "string"
            ]
          },
          "checkedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "SseItemOfFeedEntry": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FeedEntry"
          },
          "eventType": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventId": {
            "type": [
              "null",
              "string"
            ]
          },
          "reconnectionInterval": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "StandbyResponse": {
        "required": [
          "host",
          "synced",
          "confirmedFlushLsn"
        ],
        "type": "object",
        "properties": {
          "host": {
            "type": "string"
          },
          "synced": {
            "type": "boolean"
          },
          "confirmedFlushLsn": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "StatsResponse": {
        "required": [
          "at",
          "sources",
          "sinks"
        ],
        "type": "object",
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            }
          },
          "sinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SinkResponse"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Walrus.Api"
    },
    {
      "name": "Endpoints"
    }
  ]
}